pyproject.toml
[project]
name = "socratic_sofa"
version = "0.1.0"
description = "socratic-sofa using crewAI"
authors = [{ name = "Abhishek Juneja", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "anthropic>=0.75.0",
    "crewai[tools]==1.7.0",
    "gradio>=6.1.0",
    "pyyaml>=6.0.3",
    "ratelimit>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "vulture>=2.14",
    "types-PyYAML>=6.0.12",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
socratic_sofa = "socratic_sofa.main:run"
socratic_web = "socratic_sofa.gradio_app:main"
run_crew = "socratic_sofa.main:run"
train = "socratic_sofa.main:train"
replay = "socratic_sofa.main:replay"
test = "socratic_sofa.main:test"
run_with_trigger = "socratic_sofa.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

# =============================================================================
# Pre-commit Tool Configurations
# =============================================================================

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | outputs
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["socratic_sofa"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
    "outputs",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "E501",   # line too long (handled by black)
    "S101",   # use of assert (ok in tests)
    "B008",   # function call in default argument
    "S104",   # binding to all interfaces (needed for HF Spaces)
    "SIM117", # nested with statements (Gradio pattern)
    "B007",   # unused loop variable (dict iteration)
    "S110",   # try-except-pass (queue timeout pattern)
    "PTH123", # open() vs Path.open() (minor preference)
    "B904",   # raise from (style preference)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ARG"]
"src/socratic_sofa/tools/*" = ["ARG"]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "outputs"]
skips = ["B101"]  # assert used for input validation

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests", "outputs"]

[tool.vulture]
min_confidence = 80
paths = ["src/"]
exclude = ["tests/", "outputs/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
]
